Align PDF and Comic readers with shared ImageBasedReaderLayout
